webapp/requirements.txt
Django==5.1.4
python-dotenv==1.0.1
mysqlclient==2.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt
markdown==3.7
django-filter==24.3
pymupdf==1.25.1
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0